show(), product() i bundle() su sada polimorfne
kao i uslužne funkcije koje one koriste.

cilj je da su ove metode definisane u BaseController,
dok će se stvari koje su specifične (user vidi svoje recenzije
na stranici proizvoda npr) implementirati u podklasama,
da bi se izbeglo prepisivanje koda
