+++ faza5/app/Controllers/User.php
@@ -98,68 +98,6 @@ class User extends BaseController {
 
     /**
      * 
-     * Izracunavanje skora
-     * 
-     * @return double  
-     */
-    protected function getRating($positiveVotes, $negativeVotes) {
-        if ($positiveVotes == 0 && $negativeVotes == 0) return 50;
-        $totalVotes = $positiveVotes + $negativeVotes;
-        $average = $positiveVotes / $totalVotes;
-        $score = $average - ($average - 0.5) * 2 ** -log10($totalVotes + 1);
-
-        return $score * 100;
-    }
-
-    /**
-     * 
-     * Trazenje najboljih recenzija za zadati proizvod
-     * 
-     * @return array(reviews)  
-     */
-    protected function getTopReviews($id) {
-        $review_voteM = new ReviewVoteM();
-        $ownershipM = new OwnershipM();
-        $ownerships = $ownershipM->where("id_product", $id)->where("text !=", "NULL")->where("rating !=", "NULL")->findAll();
-
-        $posterScore = array();
-        $posterPosNeg = array();
-
-        foreach ($ownerships as $ownership) {
-
-            $userPoster = (new userM())->find($ownership->id_user);
-            if ($userPoster->review_ban == 1) continue;
-
-            $reviewsForPoster = $review_voteM->where('id_product', $id)->where('id_poster', $ownership->id_user)->findAll();
-            $positive = 0;
-            $negative = 0;
-
-            foreach ($reviewsForPoster as $review) {
-                if ($review->like == 0) $negative++;
-                else $positive++;
-            }
-
-            $score = $this->getRating($positive, $negative);
-            $posterScore[$ownership->id_user] = $score;
-            $posterPosNeg[$ownership->id_user] = ["positive" => $positive, "negative" => $negative];
-        }
-
-        arsort($posterScore);
-
-        $userM = new UserM();
-        $reviews = array();
-
-        foreach ($posterScore as $poster => $score) {
-            $review = $ownershipM->where('id_product', $id)->where('id_user', $poster)->first();
-            $user = $userM->find($poster);
-            $reviews[$user->username] = ["review" => $review, "positive" => $posterPosNeg[$poster]["positive"], "negative" => $posterPosNeg[$poster]["negative"]];
-        }
-
-        return $reviews;
-    }
-
-    /**
-     * 
      * Prikaz stranice proizvoda
      * 
      * @return void   
@@ -367,28 +305,6 @@ class User extends BaseController {
         }
 
         return redirect()->to(site_url("User/Product/{$id}"));
-    }
-
-
-    /**
-     * prikaži bundle sa id-jem $id
-     *
-     * @param  integer $id
-     * @return void
-     */
-    public function bundle($id) {
-        $bundle = (new BundleM())->find($id);
-
-        if (!isset($bundle))
-            return redirect()->to(site_url());
-
-        $result = [];
-        foreach ((new BundledProductsM())->findBundledProducts($id) as $idproduct) {
-            array_push($result, (new ProductM())->find($idproduct)->name);
-        }
-
-        return $this->show('bundle', ['bundle' => $bundle,
-                                      'bundledProducts' => $result]);
     }
 
 }
